refactor: update login & register components
