+++ src/app/account/index.ts
@@ -1,0 +1,2 @@
+export * from './login.component';
+export * from './register.component'
