+++ src/app/app.module.ts
@@ -16,12 +16,11 @@ import { SensorsComponent }   from './sensors';
 import { RecordsComponent }   from './records';
 import { PatientsComponent }   from './patients';
 import { ReportsComponent }   from './reports';
-import { LoginComponent }   from './login';
-import { RegisterComponent }   from './register';
+import { LoginComponent, RegisterComponent}   from './account';
+import { ProfileComponent, EditProfileComponent } from './profile';
 
 import { AlertComponent } from './_components';
 import { JwtInterceptor, ErrorInterceptor } from './_helpers';
-import { ProfileComponent, EditProfileComponent } from './profile';
 
 @NgModule({
     imports: [
