--- src/app/login/index.ts
@@ -1,1 +1,0 @@
-export * from './login.component';
