--- src/app/register/index.ts
@@ -1,1 +1,0 @@
-export * from './register.component';
